requirements.txt
# Core dependencies for cross-lingual IR experiments

# Pyserini - IR toolkit with FAISS support
pyserini>=0.22.0

# Transformers - For monoT5/mT5 reranking
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# FAISS - Dense vector search
# Choose one based on your hardware:
# For CPU:
faiss-cpu>=1.7.4
# For GPU (uncomment if you have CUDA):
# faiss-gpu>=1.7.4

# YAML configuration
PyYAML>=6.0

# Progress bars
tqdm>=4.65.0

# Optional: ColBERT support
# Uncomment if using ColBERT:
# colbert-ai>=0.2.0

# Utilities
numpy>=1.24.0

# Testing
pytest>=7.3.0

# Development dependencies (optional)
# black>=23.0.0
# flake8>=6.0.0
